pyproject.toml
[project]
name = "google-adk-ollama-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "google-adk>=1.13.0",
    "litellm>=1.76.1",
    "mkdocs-material>=9.6.18",
    "ollama>=0.5.3",
]
